requirements.txt
Django==1.8
djangorestframework==2.4.8
ipdb==0.13.3